requirements.txt
altgraph==0.17.4
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
bcrypt==4.3.0
beautifulsoup4==4.13.0b2
bleach==6.1.0
blinker==1.8.2
bs4==0.0.2
certifi==2025.4.26
cffi==1.16.0
charset-normalizer==3.3.2
chemlib==2.2.4
chempy==0.8.3
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
docutils==0.20.1
dot2tex==2.11.3
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
flask-htmx==0.4.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.47.2
fqdn==1.5.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.11
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mendeleev==0.15.0
mistune==3.0.2
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.9
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.4
opencv-python==4.10.0.82
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
pefile==2024.8.26
periodictable==1.6.1
pillow==10.2.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
PuLP==2.8.0
pure-eval==0.2.2
pycparser==2.21
pyfiglet==0.8.post1
Pygments==2.17.2
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.10.1
pymongo==4.13.0
pyneqsys==0.5.7
pyodesys==0.14.2
pyparsing==3.1.1
pypiwin32==223
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.12
PyYAML==6.0.1
pyzbar==0.1.9
pyzmq==25.1.2
quantities==0.15.0
referencing==0.32.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
scipy==1.11.4
selenium==4.27.1
Send2Trash==1.8.2
setuptools==74.1.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
sym==0.3.5
sympy==1.12
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
trio==0.28.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.2.1
xlwt==1.3.0
